Introduced command runner to separate processing from configuration
